requirements.txt
aliyun-python-sdk-core-v3==2.5.5
aliyun-python-sdk-ecs==4.6.1
amqp==2.2.2
ansible==2.2.2.0
anyjson==0.3.3
appnope==0.1.0
asn1crypto==0.23.0
Babel==2.5.3
billiard==3.5.0.3
bootstrap-admin==0.3.9
cachetools==2.0.1
celery==4.1.0
celery-with-redis==3.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
CondConfigParser==1.0.3
configparser==3.5.0
configparser2==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.2
decorator==4.1.2
Django==2.0
django-appconf==1.0.2
django-auth-ldap==1.2.16
django-bootstrap-toolkit==2.15.0
django-bootstrap3==9.1.0
django-celery==3.2.2
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-core==1.4.1
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-redis-cache==1.7.1
django-rest-swagger==2.1.2
django-simple-captcha==0.5.5
django-statici18n==1.4.0
django-tinymce==2.7.0
django-yamlfield==1.0.3
djangorestframework==3.7.1
djangorestframework-bulk==0.2.1
dotmap==1.2.20
ecdsa==0.13
enum-compat==0.0.2
ephem==3.7.6.0
eventlet==0.21.0
Flask==0.12.2
flower==0.9.2
ForgeryPy==0.1
greenlet==0.4.12
gssapi==1.2.0
idna==2.6
ipaddr==2.2.0
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
itypes==1.1.0
jedi==0.11.0
Jinja2==2.9.6
jsonfield==2.0.2
jumpserver-python-sdk==0.0.16
kombu==4.1.0
ldap3==2.3
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
npm==0.1.1
olefile==0.44
openapi-codec==1.3.2
optional-django==0.1.0
paramiko==2.1.2
parso==0.1.0
passlib==1.7.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.7
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyJSONConfigParser==0.0.1.2
pyldap==2.4.37
pypinyin==0.27.0
pyte==0.7.0
python-gssapi==0.6.4
pytz==2017.3
PyYAML==3.12
Quickstart==0.4
redis==2.10.6
requests==2.18.4
SimpleConfigParser==0.1.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
sshpubkeys==2.2.0
tornado==4.5.3
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.12.2
